# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: M&S Delivery API
  version: 1.0.0
servers:
  - url: https://api.example.com
paths:
...
  /api/v1/auth/register:
    post:
      operationId: registerCustomer
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RegisterRequest'}
      responses:
        '201':
          description: Customer registered
          content: {application/json: {schema: {$ref: '#/components/schemas/TokenPair'}}}
  /api/v1/auth/login:
    post:
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200':
          description: Authenticated
          content: {application/json: {schema: {$ref: '#/components/schemas/TokenPair'}}}
...
  /api/v1/auth/logout:
    post:
      operationId: logout
      security: [{bearerAuth: []}]
      responses:
        '204': {description: Session revoked}
  /api/v1/stores:
    get:
      operationId: listStores
      responses:
        '200': {description: Available stores}
    post:
      operationId: createStore
      security: [{bearerAuth: []}]
      responses:
        '201': {description: Store created}
  /api/v1/customers/me/addresses:
    get:
      operationId: listCustomerAddresses
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Saved addresses}
    post:
      operationId: createCustomerAddress
      security: [{bearerAuth: []}]
      responses:
        '201': {description: Address created}
  /api/v1/stores/{store_id}/menu-items:
    get:
      operationId: listMenuItems
      parameters:
        - {in: path, name: store_id, required: true, schema: {type: string, format: uuid}}
      responses:
        '200': {description: Store menu}
    post:
      operationId: createMenuItem
      security: [{bearerAuth: []}]
      parameters:
        - {in: path, name: store_id, required: true, schema: {type: string, format: uuid}}
      responses:
        '201': {description: Menu item created}
...
  /api/v1/riders:
    get:
      operationId: listRiders
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Riders}
    post:
      operationId: createRider
      security: [{bearerAuth: []}]
      responses:
        '201': {description: Rider created}
  /api/v1/riders/me/availability:
    put:
      operationId: setRiderAvailability
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Availability updated}
  /api/v1/reports/summary:
    get:
      operationId: getReportSummary
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Operational report}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  schemas:
    UserRole:
      type: string
      enum: [customer, rider, admin]
...
    RegisterRequest:
      type: object
      required: [name, email, password]
      properties:
        name: {type: string, minLength: 2}
        email: {type: string, format: email}
        phone: {type: string}
        password: {type: string, minLength: 8}
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email}
        password: {type: string}
    TokenPair:
      type: object
      required: [access_token, refresh_token, token_type, role]
      properties:
        access_token: {type: string}
        refresh_token: {type: string}
        token_type: {type: string, const: bearer}
        role: {$ref: '#/components/schemas/UserRole'}
...